Decoupling PartitionManager actor and graph entities controller methods (WIP) - Minor tuning on multithreading/asynchronous callbacks
Refers #4
