+++ mainproject/cluster/src/main/scala/com/raphtory/Actors/RaphtoryActors/PartitionManager.scala
@@ -1,6 +1,6 @@
 package com.raphtory.Actors.RaphtoryActors
 
-import java.io._
+import java.util.concurrent.atomic.AtomicInteger
 
 import akka.cluster.pubsub.{DistributedPubSub, DistributedPubSubMediator}
 import com.raphtory.GraphEntities._
@@ -9,10 +9,8 @@ import kamon.Kamon
 
 import scala.collection.concurrent.TrieMap
 import scala.concurrent.duration._
-import com.raphtory.utils.Utils._
 import monix.eval.Task
 import monix.execution.{ExecutionModel, Scheduler}
-import scala.collection.mutable
 
 /**
   * The graph partition manages a set of vertices and there edges
@@ -21,32 +19,26 @@ import scala.collection.mutable
   * */
 class PartitionManager(id : Int, test : Boolean, managerCountVal : Int) extends RaphtoryActor {
   var managerCount = managerCountVal
-  val managerID  = id                         //ID which refers to the partitions position in the graph manager map
-
-  /**
-    * Graph model maps
-    */
-  var vertices = TrieMap[Int, Vertex]()      // Map of Vertices contained in the partition
-  var edges    = TrieMap[Long, Edge]()  // Map of Edges contained in the partition
+  val managerID  = id                   //ID which refers to the partitions position in the graph manager map
 
   val printing = false                  // should the handled messages be printed to terminal
   val logging  = false                  // should the state of the vertex/edge map be output to file
 
-  var messageCount          = 0         // number of messages processed since last report to the benchmarker
-  var secondaryMessageCount = 0
-  var messageBlockID        = 0         // id of current message block to syncronise times across partition managers
+  var messageCount          : AtomicInteger = new AtomicInteger(0)         // number of messages processed since last report to the benchmarker
+  var secondaryMessageCount : AtomicInteger = new AtomicInteger(0)
   val secondaryCounting     = false     // count all messages or just main incoming ones
 
   val mediator              = DistributedPubSub(context.system).mediator // get the mediator for sending cluster messages
 
+  val storage = EntitiesStorage.apply(printing, managerCount, managerID, mediator)
+
   mediator ! DistributedPubSubMediator.Put(self)
 
-  val entitiesGauge         = Kamon.gauge("raphtory.entities")
   val verticesGauge         = Kamon.gauge("raphtory.vertices")
   val edgesGauge            = Kamon.gauge("raphtory.edges")
 
-
   implicit val s = Scheduler(ExecutionModel.BatchedExecution(1024))
+
   /**
     * Set up partition to report how many messages it has processed in the last X seconds
     */
@@ -61,27 +53,27 @@ class PartitionManager(id : Int, test : Boolean, managerCountVal : Int) extends 
 
 
   override def receive : Receive = {
-    case "tick" => reportIntake()
-    case "profile" => profile()
+    case "tick"       => reportIntake()
+    case "profile"    => profile()
     case "keep_alive" => keepAlive()
 
     //case LiveAnalysis(name,analyser) => mediator ! DistributedPubSubMediator.Send(name, Results(analyser.analyse(vertices,edges)), false)
 
-    case VertexAdd(msgId,srcId) => Task.eval(vertexAdd(msgId,srcId)).fork.runAsync; vHandle(srcId)
-    case VertexAddWithProperties(msgId,srcId,properties) => Task.eval(vertexAdd(msgId,srcId,properties)).fork.runAsync; vHandle(srcId);
-    case VertexRemoval(msgId,srcId) => Task.eval(vertexRemoval(msgId,srcId)).fork.runAsync; vHandle(srcId);
+    case VertexAdd(msgId,srcId)                                => Task.eval(storage.vertexAdd(msgId,srcId)).fork.runAsync.onComplete(_ => vHandle(srcId))
+    case VertexRemoval(msgId,srcId)                            => Task.eval(storage.vertexRemoval(msgId,srcId)).fork.runAsync.onComplete(_ => vHandle(srcId))
+    case VertexAddWithProperties(msgId,srcId,properties)       => Task.eval(storage.vertexAdd(msgId,srcId,properties)).fork.runAsync.onComplete(_ => vHandle(srcId))
 
-    case EdgeAdd(msgId,srcId,dstId) => Task.eval(edgeAdd(msgId,srcId,dstId)).fork.runAsync; eHandle(srcId,dstId)
-    case EdgeAddWithProperties(msgId,srcId,dstId,properties) => Task.eval(edgeAdd(msgId,srcId,dstId,properties)).fork.runAsync; eHandle(srcId,dstId)
-    case RemoteEdgeAdd(msgId,srcId,dstId,properties) => Task.eval(remoteEdgeAdd(msgId,srcId,dstId,properties)).fork.runAsync; eHandleSecondary(srcId,dstId)
-    case RemoteEdgeAddNew(msgId,srcId,dstId,properties,deaths) => Task.eval(remoteEdgeAddNew(msgId,srcId,dstId,properties,deaths)).fork.runAsync; eHandleSecondary(srcId,dstId)
+    case EdgeAdd(msgId,srcId,dstId)                            => Task.eval(storage.edgeAdd(msgId,srcId,dstId)).fork.runAsync.onComplete(_ => eHandle(srcId,dstId))
+    case RemoteEdgeAdd(msgId,srcId,dstId,properties)           => Task.eval(storage.remoteEdgeAdd(msgId,srcId,dstId,properties)).fork.runAsync.onComplete(_ => eHandleSecondary(srcId,dstId))
+    case RemoteEdgeAddNew(msgId,srcId,dstId,properties,deaths) => Task.eval(storage.remoteEdgeAddNew(msgId,srcId,dstId,properties,deaths)).fork.runAsync.onComplete(_ => eHandleSecondary(srcId,dstId))
+    case EdgeAddWithProperties(msgId,srcId,dstId,properties)   => Task.eval(storage.edgeAdd(msgId,srcId,dstId,properties)).fork.runAsync.onComplete(_ => eHandle(srcId,dstId))
 
-    case EdgeRemoval(msgId,srcId,dstId) => Task.eval(edgeRemoval(msgId,srcId,dstId)).fork.runAsync; eHandle(srcId,dstId)
-    case RemoteEdgeRemoval(msgId,srcId,dstId) => Task.eval(remoteEdgeRemoval(msgId,srcId,dstId)).fork.runAsync; eHandleSecondary(srcId,dstId)
-    case RemoteEdgeRemovalNew(msgId,srcId,dstId,deaths) => Task.eval(remoteEdgeRemovalNew(msgId,srcId,dstId,deaths)).fork.runAsync; eHandleSecondary(srcId,dstId)
+    case EdgeRemoval(msgId,srcId,dstId)                        => Task.eval(storage.edgeRemoval(msgId,srcId,dstId)).fork.runAsync.onComplete(_ => eHandle(srcId,dstId))
+    case RemoteEdgeRemoval(msgId,srcId,dstId)                  => Task.eval(storage.remoteEdgeRemoval(msgId,srcId,dstId)).fork.runAsync.onComplete(_ => eHandleSecondary(srcId,dstId))
+    case RemoteEdgeRemovalNew(msgId,srcId,dstId,deaths)        => Task.eval(storage.remoteEdgeRemovalNew(msgId,srcId,dstId,deaths)).fork.runAsync.onComplete(_ => eHandleSecondary(srcId,dstId))
 
-    case RemoteReturnDeaths(msgId,srcId,dstId,deaths) => Task.eval(remoteReturnDeaths(msgId,srcId,dstId,deaths)).fork.runAsync; eHandleSecondary(srcId,dstId)
-    case ReturnEdgeRemoval(msgId,srcId,dstId) => Task.eval(returnEdgeRemoval(msgId,srcId,dstId)).fork.runAsync;  eHandleSecondary(srcId,dstId)
+    case ReturnEdgeRemoval(msgId,srcId,dstId)                  => Task.eval(storage.returnEdgeRemoval(msgId,srcId,dstId)).fork.runAsync.onComplete(_ => eHandleSecondary(srcId,dstId))
+    case RemoteReturnDeaths(msgId,srcId,dstId,deaths)          => Task.eval(storage.remoteReturnDeaths(msgId,srcId,dstId,deaths)).fork.runAsync.onComplete(_ => eHandleSecondary(srcId,dstId))
 
     case UpdatedCounter(newValue) => {
       managerCount = newValue
@@ -89,241 +81,9 @@ class PartitionManager(id : Int, test : Boolean, managerCountVal : Int) extends 
     }
   }
 
-  def vertexAdd(msgId : Int, srcId : Int, properties : Map[String,String] = null) : Vertex = { //Vertex add handler function
-    var value : Vertex = new Vertex(msgId, srcId, true)
-    vertices.putIfAbsent(srcId, value) match {
-      case Some(oldValue) => {
-        oldValue revive msgId
-        value = oldValue
-      }
-      case None =>
-    }
-    if (properties != null)
-      properties.foreach(l => value + (msgId,l._1,l._2)) //add all properties
-    value
+  def keepAlive() = {
+    mediator ! DistributedPubSubMediator.Send("/user/WatchDog", PartitionUp(managerID), false)
   }
-
-  def edgeAdd(msgId : Int, srcId : Int, dstId : Int, properties : Map[String, String] = null) : Unit = {
-    val local       = checkDst(dstId, managerCount, managerID)
-    var present     = false
-    var edge : Edge = null
-    val index : Long= getEdgeIndex(srcId, dstId)
-    if (local)
-      edge = new Edge(msgId, true, srcId, dstId)
-    else
-      edge = new RemoteEdge(msgId, true, srcId, dstId, RemotePos.Destination, getPartition(dstId, managerCount))
-
-    edges.putIfAbsent(index, edge) match {
-      case Some(e) => {
-        edge = e
-        present = true
-      }
-      case None => // All is set
-    }
-    if (printing) println(s"Received an edge Add for $srcId --> $dstId (local: $local)")
-
-    if (local && srcId != dstId) {
-      val dstVertex = vertexAdd(msgId, dstId) // do the same for the destination ID
-      dstVertex addAssociatedEdge edge // do the same for the destination node
-      if (!present)
-        edge killList dstVertex.removeList
-    }
-
-    val srcVertex = vertexAdd(msgId, srcId)                          // create or revive the source ID
-    srcVertex addAssociatedEdge edge // add the edge to the associated edges of the source node
-
-    if (present) {
-        edge revive msgId
-        if (!local)
-          mediator ! DistributedPubSubMediator.Send(getManager(dstId, managerCount), RemoteEdgeAdd(msgId, srcId, dstId, null),false) // inform the partition dealing with the destination node*/
-    } else {
-        val deaths = srcVertex.removeList
-        edge killList deaths
-        if (!local)
-          mediator ! DistributedPubSubMediator.Send(getManager(dstId, managerCount), RemoteEdgeAddNew(msgId, srcId, dstId, null, deaths), false)
-    }
-    if (properties != null)
-      properties.foreach(prop => edge + (msgId,prop._1,prop._2)) // add all passed properties onto the edge
-
-  }
-
-  def remoteEdgeAdd(msgId:Int,srcId:Int,dstId:Int,properties:Map[String,String] = null):Unit={
-    if(printing) println(s"Received Remote Edge Add with properties for $srcId --> $dstId from ${getManager(srcId, managerCount)}. Edge already exists so just updating")
-    val dstVertex = vertexAdd(msgId,dstId) //create or revive the destination node
-    val edge = edges(getEdgeIndex(srcId, dstId))
-    dstVertex addAssociatedEdge edge //again I think this can be removed
-    edge revive msgId //revive  the edge
-    if (properties != null)
-      properties.foreach(prop => edge + (msgId,prop._1,prop._2)) // add all passed properties onto the list
-  }
-
-  def remoteEdgeAddNew(msgId:Int,srcId:Int,dstId:Int,properties:Map[String,String],srcDeaths:mutable.TreeMap[Int, Boolean]):Unit={
-    if(printing) println(s"Received Remote Edge Add with properties for $srcId --> $dstId from ${getManager(srcId, managerCount)}. Edge did not previously exist so sending back deaths")
-    val dstVertex = vertexAdd(msgId,dstId) //create or revive the destination node
-    val edge = new RemoteEdge(msgId,true,srcId,dstId,RemotePos.Source,getPartition(srcId, managerCount))
-    dstVertex addAssociatedEdge edge //add the edge to the associated edges of the destination node
-    edges put(getEdgeIndex(srcId,dstId), edge) //create the new edge
-    val deaths = dstVertex.removeList //get the destination node deaths
-    edge killList srcDeaths //pass source node death lists to the edge
-    edge killList deaths  // pass destination node death lists to the edge
-    properties.foreach(prop => edge + (msgId,prop._1,prop._2)) // add all passed properties onto the list
-    mediator ! DistributedPubSubMediator.Send(getManager(srcId, managerCount),RemoteReturnDeaths(msgId,srcId,dstId,deaths),false)
-  }
-
-  def getVertexAndWipe(id : Int, msgId : Int) : Vertex = {
-    vertices.get(id) match {
-      case Some(value) => value
-      case None => {
-        val x  = new Vertex(msgId,id,true)
-        vertices put(id, x)
-        x wipe()
-        x
-      }
-    }
-  }
-
-  def edgeRemoval(msgId:Int, srcId:Int, dstId:Int):Unit={
-    val local       = checkDst(dstId, managerCount, managerID)
-    var present     = false
-    var edge : Edge = null
-    val index : Long= getEdgeIndex(srcId, dstId)
-    if (local)
-      edge = new Edge(msgId, true, srcId, dstId)
-    else
-      edge = new RemoteEdge(msgId, true, srcId, dstId, RemotePos.Destination, getPartition(dstId, managerCount))
-
-    edges.putIfAbsent(index, edge) match {
-      case Some(e) => {
-        edge = e
-        present = true
-      }
-      case None => // All is set
-    }
-    if (printing) println(s"Received an edge Add for $srcId --> $dstId (local: $local)")
-    var dstVertex : Vertex = null
-    var srcVertex : Vertex = null
-
-
-    if (local && srcId != dstId) {
-      dstVertex = getVertexAndWipe(dstId, msgId)
-      dstVertex addAssociatedEdge edge // do the same for the destination node
-      if (!present)
-        edge killList dstVertex.removeList
-    }
-    srcVertex = getVertexAndWipe(srcId, msgId)
-    srcVertex addAssociatedEdge edge // add the edge to the associated edges of the source node
-
-    if (present) {
-        edge kill msgId
-        if (!local)
-          mediator ! DistributedPubSubMediator.Send(getManager(dstId, managerCount), RemoteEdgeRemoval(msgId,srcId,dstId),false) // inform the partition dealing with the destination node
-    } else {
-        val deaths = srcVertex.removeList
-        edge killList deaths
-        if (!local)
-          mediator ! DistributedPubSubMediator.Send(getManager(dstId, managerCount), RemoteEdgeRemovalNew(msgId,srcId,dstId,deaths), false)
-    }
-  }
-
-
-  def remoteEdgeRemoval(msgId:Int,srcId:Int,dstId:Int):Unit={
-    if(printing) println(s"Received Remote Edge Removal with properties for $srcId --> $dstId from ${getManager(srcId, managerCount)}. Edge already exists so just updating")
-    val dstVertex = getVertexAndWipe(dstId, msgId)
-    edges.get(getEdgeIndex(srcId, dstId)) match {
-      case Some(e) => {
-        e kill msgId
-        dstVertex addAssociatedEdge e
-      }
-      case None    => println("Didn't exist") //possibly need to fix when adding the priority box
-    }
-  }
-
-  def remoteEdgeRemovalNew(msgId:Int,srcId:Int,dstId:Int,srcDeaths:mutable.TreeMap[Int, Boolean]):Unit={
-    if(printing) println(s"Received Remote Edge Removal with properties for $srcId --> $dstId from ${getManager(srcId, managerCount)}. Edge did not previously exist so sending back deaths ")
-    val dstVertex = getVertexAndWipe(dstId, msgId)
-    val edge = new RemoteEdge(msgId,false,srcId,dstId,RemotePos.Source,getPartition(srcId, managerCount))
-    dstVertex addAssociatedEdge edge  //add the edge to the destination nodes associated list
-    edges put(getEdgeIndex(srcId,dstId), edge) // otherwise create and initialise as false
-
-    val deaths = dstVertex.removeList //get the destination node deaths
-    edge killList srcDeaths //pass source node death lists to the edge
-    edge killList deaths  // pass destination node death lists to the edge
-    mediator ! DistributedPubSubMediator.Send(getManager(srcId, managerCount),RemoteReturnDeaths(msgId,srcId,dstId,deaths),false)
-  }
-
-  def vertexRemoval(msgId:Int,srcId:Int):Unit={
-    if (printing) println(s"Received vertex remove for $srcId, updating + informing all edges")
-    var vertex : Vertex = null
-    vertices.get(srcId) match {
-      case Some(v) => {
-        vertex = v
-        v kill msgId
-      }
-      case None    => {
-        vertex = new Vertex(msgId, srcId, false)
-        vertices put (srcId, vertex)
-      }
-    }
-
-    vertex.associatedEdges.foreach(e => {
-      e kill msgId
-      if(e.isInstanceOf[RemoteEdge]){
-        val ee = e.asInstanceOf[RemoteEdge]
-        if(ee.remotePos == RemotePos.Destination) {
-          mediator ! DistributedPubSubMediator.Send(getManager(ee.remotePartitionID, managerCount), RemoteEdgeRemoval(msgId, ee.srcId, ee.dstId),false)
-        } //This is if the remote vertex (the one not handled) is the edge destination. In this case we handle with exactly the same function as above
-        else{
-          mediator ! DistributedPubSubMediator.Send(getManager(ee.remotePartitionID, managerCount), ReturnEdgeRemoval(msgId, ee.srcId, ee.dstId), false)
-        }//This is the case if the remote vertex is the source of the edge. In this case we handle it with the specialised function below
-      }
-    })
-  }
-
-  def returnEdgeRemoval(msgId:Int,srcId:Int,dstId:Int):Unit={
-    if(printing) println(s"Received Remote Edge Removal (return) for $srcId --> $dstId from ${getManager(dstId, managerCount )}. Edge already exists so just updating")
-    val srcVertex = getVertexAndWipe(srcId, msgId)
-    val edge = edges(getEdgeIndex(srcId, dstId))
-
-    srcVertex addAssociatedEdge edge //add the edge to the destination nodes associated list
-    edge kill msgId                  // if the edge already exists, kill it
-  }
-
-  def remoteReturnDeaths(msgId:Int,srcId:Int,dstId:Int,dstDeaths:mutable.TreeMap[Int, Boolean]):Unit= {
-    if(printing) println(s"Received deaths for $srcId --> $dstId from ${getManager(dstId, managerCount)}")
-    edges(getEdgeIndex(srcId,dstId)) killList dstDeaths
-  }
-
-  //***************** EDGE HELPERS
-
-
-  /********************
-   *    PRINT BLOCK   *
-   ********************/
-  def printToFile(entityName:String,msg: String):Unit={
-    val file = new File(s"/logs/${self.path.name}/entityLogs")
-    file.mkdirs()
-    val fw:FileWriter = new FileWriter(s"/logs/${self.path.name}/entityLogs/$entityName.txt")
-    try {fw.write(msg+"\n")}
-    finally fw.close()
-  }
-
-  def log(srcId:Int):Unit = if(logging && (vertices contains srcId)) printToFile(s"Vertex$srcId",vertices(srcId).printHistory())
-
-  def log(srcId:Int,dstId:Int):Unit={
-    val edgeIndex = getEdgeIndex(srcId, dstId)
-    val edge = edges(edgeIndex)
-    if(logging && (edges contains edgeIndex)) {
-      if(edge.isInstanceOf[RemoteEdge]) {
-        if(edge.asInstanceOf[RemoteEdge].remotePos == RemotePos.Source) printToFile(s"RemoteEdge$srcId-->$dstId",s"${edge.printHistory()}")
-        else printToFile(s"Edge$srcId-->$dstId",s"${edge.printHistory()}")
-      }
-      else printToFile(s"Edge$srcId-->$dstId",s"${edge.printHistory()}")
-    }
-  }
-
-  /********************
-   * .END PRINT BLOCK *
-   ********************/
 
   /*****************************
    * Metrics reporting methods *
@@ -354,41 +114,35 @@ class PartitionManager(id : Int, test : Boolean, managerCountVal : Int) extends 
   def reportIntake() : Unit = {
     if(printing)
       println(messageCount)
+
     // Kamon monitoring
-    kGauge.refine("actor" -> "PartitionManager", "name" -> "messageCount").set(messageCount)
-    kGauge.refine("actor" -> "PartitionManager", "name" -> "secondaryMessageCount").set(secondaryMessageCount)
-    reportSizes(edgesGauge, edges)
-    reportSizes(verticesGauge, vertices)
+    kGauge.refine("actor" -> "PartitionManager", "name" -> "messageCount").set(messageCount.intValue())
+    kGauge.refine("actor" -> "PartitionManager", "name" -> "secondaryMessageCount").set(secondaryMessageCount.intValue())
+
+    reportSizes(edgesGauge, storage.edges)
+    reportSizes(verticesGauge, storage.vertices)
+
     // Heap benchmarking
     //profile()
 
-    // Set counters
-    messageCount          = 0
-    secondaryMessageCount = 0
-    messageBlockID        = messageBlockID + 1
-  }
-
-  def keepAlive() = {
-    mediator ! DistributedPubSubMediator.Send("/user/WatchDog", PartitionUp(managerID), false)
+    // Reset counters
+    messageCount.set(0)
+    secondaryMessageCount.set(0)
   }
 
   def vHandle(srcID : Int) : Unit = {
-    messageCount += 1
-    kCounter.refine("actor" -> "PartitionManager", "name" -> "messageCounter").increment()
+    messageCount.incrementAndGet()
   }
 
   def vHandleSecondary(srcID : Int) : Unit = {
-    secondaryMessageCount += 1
-    kCounter.refine("actor" -> "PartitionManager", "name" -> "secondaryMessageCounter").increment()
+    secondaryMessageCount.incrementAndGet()
   }
   def eHandle(srcID : Int, dstID : Int) : Unit = {
-    messageCount += 1
-    kCounter.refine("actor" -> "PartitionManager", "name" -> "messageCounter").increment()
+    messageCount.incrementAndGet()
   }
 
   def eHandleSecondary(srcID : Int, dstID : Int) : Unit = {
-    secondaryMessageCount += 1
-    kCounter.refine("actor" -> "PartitionManager", "name" -> "secondaryMessageCounter").increment()
+    secondaryMessageCount.incrementAndGet()
   }
 
   /*********************************
