+++ mainproject/cluster/src/main/scala/com/raphtory/Actors/RaphtoryActors/UpdateGen.scala
@@ -217,5 +217,4 @@ class UpdateGen extends RaphtoryActor with Timers {
     }
     properties
   }
-
 }
